refactor: use methods for reference access
